Intel_AI_Processor/requirements.txt
streamlit==1.30.0
fastapi==0.109.0
uvicorn==0.27.1
langchain-community==0.3.19
openai==1.8.0
pinecone-client==3.0.0
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.31.0
tiktoken==0.6.0
pypdf==4.0.1
